change some variables to 16 and 8 bit width integers, implement remaining move types (capture, castling, promotion, etc.), add graveyard and new movement control code (untested, currently commented out)

+++ Src/include/queue.h
@@ -1,6 +1,8 @@
 #ifndef __QUEUE_H_
 #define __QUEUE_H_
 
+#include <stdint.h>
+
 #define SIZE 128 // proabably needs to be bigger
 #define INCREMENT(q,s) (q)->s = ((q)->s == &(q->queue[SIZE-1])) ? (q)->queue : (q)->s+1
 
@@ -8,8 +10,8 @@ typedef int (*done_func)(void);
 
 // steps tuple
 typedef struct {
-	int x;
-	int y;
+	int16_t x;
+	int16_t y;
 	done_func done; // function to run when done
 } tuple_t;
 
@@ -21,7 +23,7 @@ typedef struct {
 } tuple_queue_t;
 
 void init(tuple_queue_t *q);
-unsigned char add(tuple_queue_t *q, int x, int y, done_func d);
+unsigned char add(tuple_queue_t *q, int16_t x, int16_t y, done_func d);
 unsigned char is_empty(tuple_queue_t *q);
 unsigned char is_full(tuple_queue_t *q);
 void clear_queue(tuple_queue_t *q);
